# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk mencatat servis kendaraan dan manajemen foto per transaksi.",
        "title": "Service Kendaraan API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/bengkel": {
            "get": {
                "description": "Mengambil data bengkel. Contoh: ?page=1\u0026limit=5 atau ?page=2\u0026limit=15",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master Bengkel"
                ],
                "summary": "List Semua Bengkel dengan pagination fleksibel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter Aktif (true, false, all)",
                        "name": "aktif",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Tipe Bengkel (mobil bbm, motor bbm, mobil ev, listrik ev, mobil bbm \u0026 ev, motor bbm \u0026 ev)",
                        "name": "tipe",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Halaman ke berapa (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Jumlah data per halaman (bisa 5, 10, 15, dll)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "nama",
                        "description": "Kolom untuk sorting (nama, jenis_service, tipe, created_at)",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "asc",
                        "description": "Arah urutan (asc, desc)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Cari berdasarkan nama bengkel",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter nama Provinsi",
                        "name": "provinsi",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter nama Kota/Kabupaten",
                        "name": "kota_kab",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter nama Kecamatan",
                        "name": "kecamatan",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter nama Kelurahan",
                        "name": "kelurahan",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.PaginationResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Bengkel"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master Bengkel"
                ],
                "summary": "Tambah Bengkel Baru",
                "parameters": [
                    {
                        "description": "Data Bengkel",
                        "name": "bengkel",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.BengkelDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Bengkel"
                        }
                    }
                }
            }
        },
        "/api/bengkel/getBengkelByStatus/{status}": {
            "get": {
                "description": "Mengambil data Bengkel berdasarkan status. Contoh: /api/bengkel/getBengkelByStatus/",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master Bengkel"
                ],
                "summary": "Ambil Data Bengkel berdasarkan status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Status",
                        "name": "status",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Bengkel"
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/bengkel/{id}": {
            "get": {
                "description": "Mengambil data spesifik bengkel berdasarkan ID. Contoh: /api/bengkel/1",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master Bengkel"
                ],
                "summary": "Ambil Data Bengkel berdasarkan ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Bengkel",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data",
                        "schema": {
                            "$ref": "#/definitions/models.Bengkel"
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Mengubah data bengkel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master Bengkel"
                ],
                "summary": "Update Data Bengkel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Bengkel",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data Bengkel",
                        "name": "bengkel",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Bengkel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Bengkel"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.BengkelDTO": {
            "type": "object",
            "required": [
                "alamat",
                "jenis_service",
                "kecamatan",
                "kelurahan",
                "kota_kab",
                "nama",
                "provinsi",
                "tipe"
            ],
            "properties": {
                "alamat": {
                    "type": "string"
                },
                "jenis_service": {
                    "type": "string"
                },
                "kecamatan": {
                    "type": "string"
                },
                "kelurahan": {
                    "type": "string"
                },
                "kota_kab": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "no_telp": {
                    "type": "string"
                },
                "provinsi": {
                    "type": "string"
                },
                "tipe": {
                    "type": "string"
                }
            }
        },
...
        "models.Bengkel": {
            "type": "object",
            "properties": {
                "aktif": {
                    "type": "boolean"
                },
                "alamat": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "jenis_service": {
                    "type": "string"
                },
                "kecamatan": {
                    "type": "string"
                },
                "kelurahan": {
                    "type": "string"
                },
                "kota_kab": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "no_telp": {
                    "type": "string"
                },
                "provinsi": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "example": "MOBIL BBM"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "utils.PaginationResult": {
            "type": "object",
            "properties": {
                "data": {},
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total_data": {
                    "type": "integer",
                    "example": 100
                },
                "total_page": {
                    "type": "integer",
                    "example": 10
                }
            }
        }
    }
}